requirements.txt
# Core
anthropic>=0.40.0
gradio>=4.44.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0

# Telegram
python-telegram-bot>=21.0.0

# Supabase
supabase>=2.9.0

# Data & ML
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
statsmodels>=0.14.0
openpyxl>=3.1.0        # Lectura de archivos Excel (futuro: datos reales del negocio)
xlrd>=2.0.1

# HTTP
httpx>=0.27.0
aiohttp>=3.10.0

# Utilidades
python-dateutil>=2.9.0
pydantic>=2.9.0
